restaurants_api/requirements.txt
djangorestframework
markdown
djangorestframework-api-key==2.0.0
